ml_service/requirements.txt
flask
tensorflow==2.20.0
numpy
Pillow
